parcial2/requirements.txt
flask
PyYAML